move plotly related tests to test_plotly_backend
