+++ nilearn/plotting/surface/tests/test_surf_plotting.py
@@ -23,9 +23,6 @@ from nilearn.plotting import (
     plot_surf_stat_map,
 )
 from nilearn.plotting.displays import PlotlySurfaceFigure, SurfaceFigure
-from nilearn.plotting.surface._plotly_backend import (
-    _get_view_plot_surf_plotly,
-)
 from nilearn.plotting.surface._utils import check_surface_plotting_inputs
 from nilearn.surface import (
     InMemoryMesh,
@@ -38,140 +35,6 @@ except ImportError:
     IPYTHON_INSTALLED = False
 else:
     IPYTHON_INSTALLED = True
-
-
-EXPECTED_CAMERAS_PLOTLY = [
-    (
-        "left",
-        "lateral",
-        (0, 180),
-        {
-            "eye": {"x": -1.5, "y": 0, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    (
-        "left",
-        "medial",
-        (0, 0),
-        {
-            "eye": {"x": 1.5, "y": 0, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Dorsal left
-    (
-        "left",
-        "dorsal",
-        (90, 0),
-        {
-            "eye": {"x": 0, "y": 0, "z": 1.5},
-            "up": {"x": -1, "y": 0, "z": 0},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Ventral left
-    (
-        "left",
-        "ventral",
-        (270, 0),
-        {
-            "eye": {"x": 0, "y": 0, "z": -1.5},
-            "up": {"x": 1, "y": 0, "z": 0},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Anterior left
-    (
-        "left",
-        "anterior",
-        (0, 90),
-        {
-            "eye": {"x": 0, "y": 1.5, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Posterior left
-    (
-        "left",
-        "posterior",
-        (0, 270),
-        {
-            "eye": {"x": 0, "y": -1.5, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Lateral right
-    (
-        "right",
-        "lateral",
-        (0, 0),
-        {
-            "eye": {"x": 1.5, "y": 0, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Medial right
-    (
-        "right",
-        "medial",
-        (0, 180),
-        {
-            "eye": {"x": -1.5, "y": 0, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Dorsal right
-    (
-        "right",
-        "dorsal",
-        (90, 0),
-        {
-            "eye": {"x": 0, "y": 0, "z": 1.5},
-            "up": {"x": -1, "y": 0, "z": 0},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Ventral right
-    (
-        "right",
-        "ventral",
-        (270, 0),
-        {
-            "eye": {"x": 0, "y": 0, "z": -1.5},
-            "up": {"x": 1, "y": 0, "z": 0},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Anterior right
-    (
-        "right",
-        "anterior",
-        (0, 90),
-        {
-            "eye": {"x": 0, "y": 1.5, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-    # Posterior right
-    (
-        "right",
-        "posterior",
-        (0, 270),
-        {
-            "eye": {"x": 0, "y": -1.5, "z": 0},
-            "up": {"x": 0, "y": 0, "z": 1},
-            "center": {"x": 0, "y": 0, "z": 0},
-        },
-    ),
-]
 
 
 @pytest.fixture
@@ -277,38 +140,6 @@ def test_plot_surf_contours_warning_hemi(in_memory_mesh):
     parcellation[in_memory_mesh.faces[3]] = 1
     with pytest.warns(UserWarning, match="This value will be ignored"):
         plot_surf_contours(in_memory_mesh, parcellation, hemi="left")
-
-
-@pytest.mark.parametrize("full_view", EXPECTED_CAMERAS_PLOTLY)
-def test_get_view_plot_surf_plotly(full_view):
-    from nilearn.plotting.surface._plotly_backend import (
-        _get_camera_view_from_elevation_and_azimut,
-        _get_camera_view_from_string_view,
-    )
-
-    hemi, view_name, (elev, azim), expected_camera_view = full_view
-    camera_view = _get_view_plot_surf_plotly(hemi, view_name)
-    camera_view_string = _get_camera_view_from_string_view(hemi, view_name)
-    camera_view_elev_azim = _get_camera_view_from_elevation_and_azimut(
-        (elev, azim)
-    )
-    # Check each camera view parameter
-    for k in ["center", "eye", "up"]:
-        # Check default camera view
-        assert np.allclose(
-            list(camera_view[k].values()),
-            list(expected_camera_view[k].values()),
-        )
-        # Check camera view obtained from string view
-        assert np.allclose(
-            list(camera_view_string[k].values()),
-            list(expected_camera_view[k].values()),
-        )
-        # Check camera view obtained from elevation & azimut
-        assert np.allclose(
-            list(camera_view_elev_azim[k].values()),
-            list(expected_camera_view[k].values()),
-        )
 
 
 def test_surface_figure():
@@ -583,42 +414,6 @@ def test_add_contours_line_properties(plotly, key, value, surface_image_roi):
     figure.add_contours(surface_image_roi, levels=[1], lines=[{key: value}])
     newline = figure.figure.to_dict().get("data")[2].get("line")
     assert newline.get(key) == value
-
-
-@pytest.mark.parametrize("hemi,view", [("foo", "medial"), ("bar", "anterior")])
-def test_get_view_plot_surf_hemisphere_errors(hemi, view):
-    with pytest.raises(ValueError, match="Invalid hemispheres definition"):
-        _get_view_plot_surf_plotly(hemi, view)
-
-
-@pytest.mark.parametrize(
-    "hemi,view,f",
-    [
-        ("left", "foo", _get_view_plot_surf_plotly),
-        ("right", "bar", _get_view_plot_surf_plotly),
-        ("both", "lateral", _get_view_plot_surf_plotly),
-        ("both", "medial", _get_view_plot_surf_plotly),
-        ("both", "foo", _get_view_plot_surf_plotly),
-    ],
-)
-def test_get_view_plot_surf_view_errors(hemi, view, f):
-    with pytest.raises(ValueError, match="Invalid view definition"):
-        f(hemi, view)
-
-
-def test_configure_title_plotly():
-    from nilearn.plotting.surface.surf_plotting import _configure_title_plotly
-
-    assert _configure_title_plotly(None, None) == {}
-    assert _configure_title_plotly(None, 22) == {}
-    config = _configure_title_plotly("Test Title", 22, color="green")
-    assert config["text"] == "Test Title"
-    assert config["x"] == 0.5
-    assert config["y"] == 0.96
-    assert config["xanchor"] == "center"
-    assert config["yanchor"] == "top"
-    assert config["font"]["size"] == 22
-    assert config["font"]["color"] == "green"
 
 
 def test_plot_surf_engine_error(in_memory_mesh):
